+++ src/swipl-devel/src/pl-vmi.c
@@ -503,33 +503,22 @@ END_VMI
 VMH(h_const, 1, (word), (c))
 { Word k = ARGP;
 
-  for(;;)
-  { switch(tag(*k))
-    { case TAG_VAR:
-	if ( !hasTrailSpace(1) )
-	  break;
-	varBindConst(k, c);
-	ARGP++;
-	NEXT_INSTRUCTION;
-      case TAG_ATTVAR:
-	if ( !hasGlobalSpace(0) )
-	  break;
-	assignAttVar(k, &c);
-	ARGP++;
-	NEXT_INSTRUCTION;
-      case TAG_REFERENCE:
-	k = unRef(*k);
-	continue;
-      default:
-	if ( *k == c )
-	{ ARGP++;
-	  NEXT_INSTRUCTION;
-	}
-	CLAUSE_FAILED;
-    }
-    GROW_STACK_SPACE(0,0);
-    k = ARGP;
+  deRef(k);
+  if ( *k == c )
+  { ARGP++;
+    NEXT_INSTRUCTION;
   }
+  if ( canBind(*k) )
+  { if ( !hasGlobalSpace(0) )
+    { GROW_STACK_SPACE(0,0);
+      k = ARGP;
+      deRef(k);
+    }
+    bindConst(k, c);
+    ARGP++;
+    NEXT_INSTRUCTION;
+  }
+  CLAUSE_FAILED;
 }
 END_VMH
 
@@ -686,7 +675,7 @@ stack that must be treated as a variable.
 - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 
 VMI(H_VAR, 0, 1, (CA1_VAR))
-{ Word k = varFrameP(FR, (int)*PC++);
+{ Word k = varFrameP(FR, (size_t)*PC++);
   int rc;
 
   if ( UMODE == uwrite )
@@ -723,6 +712,20 @@ VMI(H_VAR, 0, 1, (CA1_VAR))
     } else
     { setVar(*ARGP);
     }
+  }
+
+  /* First try the simple case.  do_unify() either completes
+   * (TRUE or FALSE) or returns a memory overflow code.  In
+   * the latter case we just try again, protected for GC
+   */
+  if ( LD->prolog_flag.occurs_check == OCCURS_CHECK_FALSE )
+  { int rc = do_unify(k, ARGP);
+    if ( rc == TRUE )
+    { ARGP++;
+      NEXT_INSTRUCTION;
+    }
+    if ( rc == FALSE )
+      CLAUSE_FAILED;
   }
 
   SAVE_REGISTERS(QID);
@@ -2027,7 +2030,7 @@ VMH(depart_or_retry_continue, 0, (), ())
       }
     }
 #endif /*O_DEBUGGER*/
-  }
+  } /* end of if (LD->alerted) */
 
   PC = DEF->codes;
   NEXT_INSTRUCTION;
@@ -2442,8 +2445,8 @@ simplified version of linkVal().
 - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 
 VMI(L_VAR, 0, 2, (CA1_FVAR,CA1_VAR))
-{ Word v1 = varFrameP(FR, (int)*PC++);
-  Word v2 = varFrameP(FR, (int)*PC++);
+{ Word v1 = varFrameP(FR, (size_t)*PC++);
+  Word v2 = varFrameP(FR, (size_t)*PC++);
   word w = *v2;
 
   while(isRef(w))
@@ -2459,7 +2462,7 @@ VMI(L_VAR, 0, 2, (CA1_FVAR,CA1_VAR))
 END_VMI
 
 VMI(L_VOID, 0, 1, (CA1_FVAR))
-{ Word v1 = varFrameP(FR, (int)*PC++);
+{ Word v1 = varFrameP(FR, (size_t)*PC++);
 
   setVar(*v1);
   NEXT_INSTRUCTION;
@@ -2467,7 +2470,7 @@ VMI(L_VOID, 0, 1, (CA1_FVAR))
 END_VMI
 
 VMI(L_ATOM, 0, 2, (CA1_FVAR,CA1_DATA))
-{ Word v1 = varFrameP(FR, (int)*PC++);
+{ Word v1 = varFrameP(FR, (size_t)*PC++);
   word  c = code2atom(*PC++);
   pushVolatileAtom(word2atom(c));
   *v1 = c;
@@ -2476,7 +2479,7 @@ VMI(L_ATOM, 0, 2, (CA1_FVAR,CA1_DATA))
 END_VMI
 
 VMI(L_NIL, 0, 1, (CA1_FVAR))
-{ Word v1 = varFrameP(FR, (int)*PC++);
+{ Word v1 = varFrameP(FR, (size_t)*PC++);
 
   *v1 = ATOM_nil;
   NEXT_INSTRUCTION;
@@ -4617,6 +4620,7 @@ VMH(foreign_redo, 0, (), ())
 { Choice ch;
   FliFrame ffr;
 
+  environment_frame = FR;
   lTop = (LocalFrame)argFrameP(FR, DEF->functor->arity);
   ch = newChoice(CHP_JUMP, FR);
   ch->value.pc = PC+3;
@@ -4624,6 +4628,7 @@ VMH(foreign_redo, 0, (), ())
   ffr = (FliFrame)(ch+1);
   lTop = (LocalFrame)(ffr+1);
   ffr->size = 0;
+  ffr->no_free_before = (size_t)-1;
   NoMark(ffr->mark);
   ffr->parent = fli_context;
   FLI_SET_VALID(ffr);
@@ -5090,7 +5095,8 @@ again:
 	    Sdprintf(" (retrying)\n"));
 
       PL_rewind_foreign_frame(fid);
-      clear((LocalFrame)valTermRef(catchfr_ref), FR_CATCHED);
+      if ( catchfr_ref )
+	clear((LocalFrame)valTermRef(catchfr_ref), FR_CATCHED);
       goto again;
     }
   }
@@ -5958,18 +5964,21 @@ VMI(S_TRIE_GEN, 0, 0, ())
     { if ( t->value_count == 0 )
 	FRAME_FAILED;
 
+      LD->query->next_environment = lTop;
+      lTop = (LocalFrame)ARGP;
       SAVE_REGISTERS(QID);
       dbref = compile_trie(FR->predicate, t);
       LOAD_REGISTERS(QID);
+      LD->query->next_environment = NULL;
       if ( dbref == ATOM_error )
 	THROW_EXCEPTION;
     }
 
     if ( dbref == ATOM_fail )
       FRAME_FAILED;
-  }
 
-  cref = clause_clref(dbref);
+    cref = clause_clref(dbref);
+  }
 
   ARGP = argFrameP(FR, 0);
   TRUST_CLAUSE(cref);
@@ -6321,9 +6330,9 @@ VMI(T_FLOAT, 0, CODES_PER_DOUBLE, (CA1_FLOAT))
     TrieNextArg();
     NEXT_INSTRUCTION;
   } else if ( isFloat(*k) )
-  { Word p = valIndirectP(*k);
+  { Code p = (Code)valIndirectP(*k);
 
-    switch(WORDS_PER_DOUBLE) /* depend on compiler to clean up */
+    switch(CODES_PER_DOUBLE) /* depend on compiler to clean up */
     { case 2:
 	if ( *p++ != *PC++ )
 	  CLAUSE_FAILED;
@@ -6415,13 +6424,16 @@ END_VMI
 
 VMI(T_TRY_SMALLINTW, 0, 1+CODES_PER_WORD, (CA1_JUMP,CA1_WORD))
 { TRIE_TRY;
-  VMI_GOTO(T_SMALLINT);
+  VMI_GOTO(T_SMALLINTW);
 }
 END_VMI
 
 VMI(T_SMALLINTW, 0, CODES_PER_WORD, (CA1_WORD))
 { word w;
 
+#if CODES_PER_WORD == 1
+  SEPARATE_VMI1;		/* Might collapse with T_SMALLINT */
+#endif
   PC = code_get_word(PC, &w);
   DEBUG(MSG_TRIE_VM, Sdprintf("T_SMALLINT %lld\n", (long long)w));
   VMH_GOTO(t_const, consInt((sword)w));