+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(rswipl)
-
-test_check("rswipl")